+++ src/app/GitUI/LeftPanel/FavoriteNode.cs
@@ -3,6 +3,7 @@
 internal class FavoriteNode : BaseRevisionNode
 {
     private readonly string _imageKey;
+
     public FavoriteNode(FavoritesTree tree, string name, string imageKey)
         : base(tree, name, true)
     {
